Former-commit-id: 8c4239aa407677acce6f8f744842b805e8ad4f8e

+++ dotCMS/html/portlet/ext/useradmin/view_users_js_inc.jsp
@@ -376,6 +376,13 @@
 		userChanged = true;
 		passwordChanged = true;
 	}
+	
+	//Handler from when the user info has changed
+	var emailChanged = false;
+	function userEmailChanged() {
+		userChanged = true;
+		emailChanged = true;
+	}
 
 	//Handler to save the user details
 	function saveUserDetails() {
@@ -393,7 +400,7 @@
 		}
 
 		var userEmail;
-		if(currentUser == null){
+		if(emailChanged){
 			var pattern=/^([a-zA-Z0-9_.-])+@([a-zA-Z0-9_.-])+\.([a-zA-Z])+([a-zA-Z])+/;
 			userEmail = dijit.byId('emailAddress').attr('value');
 			if(!pattern.test(userEmail)){
@@ -420,7 +427,7 @@
 		}
 		if(!newUser) {
 			UserAjax.updateUser(currentUser.id, currentUser.id, dijit.byId('firstName').attr('value'), dijit.byId('lastName').attr('value'),
-					dijit.byId('emailAddress').attr('value'), passswordValue, callbackOptions);
+					userEmail, passswordValue, callbackOptions);
 		} else {
 			if (!authByEmail) {
 				UserAjax.addUser(dijit.byId('userId').attr('value'), dijit.byId('firstName').attr('value'), dijit.byId('lastName').attr('value'), dijit.byId('emailAddress').attr('value'), passswordValue, callbackOptions);
@@ -967,7 +974,8 @@
 
 	//Saves the current selection of roles
 	function saveRoles () {
-		UserAjax.updateUserRoles(currentUser.id, rolesAdded, saveRolesCallback);
+		if(currentUser != null)
+			UserAjax.updateUserRoles(currentUser.id, rolesAdded, saveRolesCallback);
 	}
 
 	//Callback from the server after successful save
@@ -1202,9 +1210,10 @@
 	}
 
 	function saveAddress(){
-		if(!dijit.byId('addressForm').validate())
+		if(currentUser == null)
 			return;
-
+		if(!dijit.byId('addressForm').validate())
+			return;		
 		var id = dojo.byId('addressId').value;
 		var desc = dijit.byId('addressDescription').attr('value');
 		var street1 = dijit.byId('addressStreet1').attr('value');
@@ -1286,9 +1295,10 @@
 	}
 
 	function saveUserAdditionalInfo(){
-		if (!dijit.byId('userAdditionalInfoForm').validate())
+		if(currentUser == null)
 			return;
-
+		if (!dijit.byId('userAdditionalInfoForm').validate())
+			return;		
 		var active = dijit.byId('userActive').attr('value') != false;
 		var prefix = dijit.byId('prefix').attr('value');
 		var suffix = dijit.byId('suffix').attr('value');
@@ -1356,6 +1366,8 @@
 	}
 
 	function viewFullClickHistory() {
+		if(currentUser == null)
+			return;
 		if(dijit.byId('userClickHistoryPane'))
 			dijit.registry.remove('userClickHistoryPane');
 
@@ -1379,6 +1391,8 @@
 	}
 
 	function viewClickstreamDetails(clickstreamId, userId) {
+		if(currentUser == null)
+			return;
 		if(dijit.byId('userClickHistoryDetailPane'))
 			dijit.registry.remove('userClickHistoryDetailPane');
 
@@ -1400,10 +1414,12 @@
 	//User Tags
 
 	function initTags() {
-		TagAjax.getTagsByUser(currentUser.userId, showResult);
+		if(currentUser != null)
+			TagAjax.getTagsByUser(currentUser.userId, showResult);
 	}
 	function removeTagInode(tagName) {
-		TagAjax.deleteTag(tagName, currentUser.userId, showResult);
+		if(currentUser != null)
+			TagAjax.deleteTag(tagName, currentUser.userId, showResult);
 	}
 	function editTag(tagName) {
 		var tagTable = document.getElementById('tags_detail');
@@ -1418,7 +1434,8 @@
 		var cmd = document.getElementById('cmd').value;
 		document.getElementById('tagName').value = '';
 		document.getElementById('cmd').value = '';
-		TagAjax.addTag(tagName, currentUser.userId, currentUser.inode, showResult);
+		if(currentUser != null)
+			TagAjax.addTag(tagName, currentUser.userId, currentUser.inode, showResult);
 	}
 	function showResult(result) {
 		DWRUtil.removeAllRows("tags_table");
@@ -1461,7 +1478,8 @@
 
 	//User categories
 	function loadUserCategories() {
-		UserAjax.getUserCategories(currentUser.userId, loadUserCategoriesCallback);
+		if(currentUser != null)
+			UserAjax.getUserCategories(currentUser.userId, loadUserCategoriesCallback);
 	}
 
 	function loadUserCategoriesCallback(categories) {
@@ -1484,7 +1502,8 @@
 				}
 			}
 		}, this);
-		UserAjax.updateUserCategories(currentUser.userId, selectedCategories, updateUserCategoriesCallback);
+		if(currentUser != null)
+			UserAjax.updateUserCategories(currentUser.userId, selectedCategories, updateUserCategoriesCallback);
 	}
 
 	function updateUserCategoriesCallback(){
@@ -1527,8 +1546,8 @@
 	function updateUserLocale() {
 		var timeZoneId = dijit.byId('userTimeZone').attr('value');
 		var languageId = dijit.byId('userLanguage').attr('value');
-
-		UserAjax.updateUserLocale(currentUser.userId, timeZoneId, languageId, updateUserLocaleCallback);
+		if(currentUser != null)
+			UserAjax.updateUserLocale(currentUser.userId, timeZoneId, languageId, updateUserLocaleCallback);
 	}
